Links Browser interface

Uses the Links Text Browser to read a web page.

Usage

read.links(url, intern = TRUE, bin = NULL, pipe = FALSE, ...)

Arguments

url

a character string specifying the URL of the web page.

intern

a logical which indicates whether to make the output of the command an R object.

bin

a string with the path of your lynx binary or NULL if lynx binary is available in the operating system path.

pipe

a logical which indicates whether the result should be returned as a pipe() commmand.

...

optional arguments passed to links binary.

Value

the downloaded text

References

Diethelm Wuertz, Yohan Chalabi, and Andrew Ellis, (2010); Financial Market Data for R/Rmetrics, Rmetrics eBook, Rmetrics Association and Finance Online, Zurich, www.rmetrics.org.

See also